Split CellError
