+++ auraed/src/runtime/cells/cells_table.rs
@@ -28,7 +28,8 @@
  *                                                                            *
 \* -------------------------------------------------------------------------- */
 
-use crate::runtime::cells::error::CellError;
+use crate::runtime::cells::cell::CellError;
+use crate::runtime::cells::CellServiceError;
 use crate::runtime::cells::{Cell, CellName, Result};
 use std::{
     collections::HashMap,
@@ -54,15 +55,17 @@ impl CellsTable {
     /// The clone can be retrieved once it's removed from the cache.
     /// Returns an error if a duplicate [cell_name] already exists in the cache.
     pub(crate) fn insert(&self, cell_name: CellName, cell: Cell) -> Result<()> {
-        let mut cache =
-            self.cache.lock().map_err(|_| CellError::FailedToObtainLock())?;
+        let mut cache = self
+            .cache
+            .lock()
+            .map_err(|_| CellServiceError::FailedToObtainLock())?;
 
         // TODO: replace with this when it becomes stable
         // cache.try_insert(cell_name.clone(), cgroup)
 
         // Check if there was already a cgroup in the table with this cell name as a key.
         if cache.contains_key(&cell_name) {
-            return Err(CellError::Exists { cell_name: cell_name.clone() });
+            return Err(CellError::Exists { cell_name }.into());
         }
         // Ignoring return value as we've already assured ourselves that the key does not exist.
         let _ = cache.insert(cell_name, cell);
@@ -70,8 +73,10 @@ impl CellsTable {
     }
 
     pub(crate) fn contains(&self, cell_name: &CellName) -> Result<bool> {
-        let cache =
-            self.cache.lock().map_err(|_| CellError::FailedToObtainLock())?;
+        let cache = self
+            .cache
+            .lock()
+            .map_err(|_| CellServiceError::FailedToObtainLock())?;
 
         Ok(cache.contains_key(cell_name))
     }
@@ -92,25 +97,29 @@ impl CellsTable {
     where
         F: FnOnce(&mut Cell) -> Result<R>,
     {
-        let mut cache =
-            self.cache.lock().map_err(|_| CellError::FailedToObtainLock())?;
+        let mut cache = self
+            .cache
+            .lock()
+            .map_err(|_| CellServiceError::FailedToObtainLock())?;
 
         if let Some(cell) = cache.get_mut(cell_name) {
             f(cell)
         } else {
-            Err(CellError::NotFound { cell_name: cell_name.clone() })
+            Err(CellError::NotFound { cell_name: cell_name.clone() }.into())
         }
     }
 
     /// Remove and return the cgroup keyed by [cell_name] from the cache.
     /// Returns an error if the cell_name does not exist in the cache.
     pub(crate) fn remove(&self, cell_name: &CellName) -> Result<Cell> {
-        let mut cache =
-            self.cache.lock().map_err(|_| CellError::FailedToObtainLock())?;
+        let mut cache = self
+            .cache
+            .lock()
+            .map_err(|_| CellServiceError::FailedToObtainLock())?;
 
-        cache
-            .remove(cell_name)
-            .ok_or_else(|| CellError::NotFound { cell_name: cell_name.clone() })
+        cache.remove(cell_name).ok_or_else(|| {
+            CellError::NotFound { cell_name: cell_name.clone() }.into()
+        })
     }
 }
